+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add support for custom request headers
 - Move everything to a single class
 - Add private build_context() method
+- List of default request headers, currently user_agent and accept
 
 ## 0.0.1 - 2023-04-06
 - Basic minimal working version